ai-server/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.9.2
pymongo==4.10.1
openai==1.59.7
python-dotenv==1.0.1

scikit-learn
numpy
